Possible solution for the symbolic link problem
Construct a temporary 'bridge' alias that pairs the symbolic link's
parent with the target node.  This returns an Alias instance that points
to the right node, has the right name and has the right parent.  Need to
see how actual applications handle things like .. to know if this will
suffice.