In the middle of another massive refactor
Getting rid of the entry point vectors and all the nonsense that was
associated with them.  Using the default RPC entry point vector should
be fine since context handles are being used; pretty sure the same UUID
can also be used for both objects but haven't tested it yet.  This is a
milestone check-in, there is much more to do for this update