Allow MemorySection to manage a length granularity
This helps when multiple sections are intended to appear contiguous to
the running process.  For example, specifying a 64K length granularity
prevents unusable address space holes from appearing, as the next
available address will be properly aligned.  By default, the class will
use the system page size for this adjustment.