Fix problems with OutputDebugString / _RPTx macros in host
The underlying cause is trying to use OutputDebugString from a vectored
exception handler.  Deal with this by catching the debug exception and
sending the string to the service over RPC instead.  This will
ultimately work out better regardless, there will be much more
flexibility dealing with this on the server side.