Only remove non-orphaned Threads on context rundown
This at least partially solves a problem where an orphaned thread, like
what would occur from an execve(), would send the stale TID into the
parent Process for removal.  If this were to occur while another active
Thread with the same TID is in the Process, that active thread would be
removed from the Process instead.