Flesh out new Session, ProcessGroup and Namespace classes a bit
The chicken-and-egg problem for these classes was solved by accessing
the member collection in the static factory function.  This model will
also propagate to Process and Thread, there really is no other clean way
to do this, the child needs the parent and the parent cannot exist
without an initial child.